fix: Home controlls chat store and sessions
